requirements.txt
asgiref                       3.7.2
Django                        5.0
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.1
pip                           23.3.2
PyJWT                         2.8.0
pytz                          2023.3.post1
sqlparse                      0.4.4
tzdata                        2023.3